fix: rename PascalCase module files to camelCase per AGENTS.md convention
PostgresCandleReadAdapter → postgresCandleReadAdapter
PostgresCandleRevisionUnitOfWork → postgresCandleRevisionUnitOfWork
SqliteCandleReadAdapter → sqliteCandleReadAdapter
SqliteCandleRevisionUnitOfWork → sqliteCandleRevisionUnitOfWork
IngestCandlesUseCase → ingestCandlesUseCase
FakeCandleWritePort → fakeCandleWritePort
