+++ src/adapters/sqlite/sqliteCandleReadAdapter.ts
