+++ src/application/use-cases/__tests__/fakes/fakeCandleWritePort.ts
