+++ src/application/use-cases/ingestCandlesUseCase.ts
