+++ src/http/handlers/candlesIngest.ts
@@ -1,7 +1,7 @@
 import type { FastifyReply, FastifyRequest } from "fastify";
 import { SCHEMA_VERSION, type CandleIngestResponse } from "../../contract/v1/types.js";
 import { parseCandleIngestRequest } from "../../contract/v1/validation.js";
-import type { IngestCandlesUseCase } from "../../application/use-cases/IngestCandlesUseCase.js";
+import type { IngestCandlesUseCase } from "../../application/use-cases/ingestCandlesUseCase.js";
 import type { ClockPort } from "../../application/ports/clock.js";
 import { AuthError, requireSharedSecret } from "../auth.js";
 import { ContractValidationError } from "../errors.js";
